Debug githubaction buildocker (#9)
Running of github action to build docker image only allowed in a branch not in main, so versions file can be updated and pushed. So after that make a pull request and approve (by admins) new version of KEIT docker image. Bump version to 0.0.6. This pull request does it also so now we have a new version. (which is already in ghcr).